All works, all testers pass, need to check leftovers

+++ message_slot.h
@@ -4,7 +4,7 @@
 #include <linux/ioctl.h>
 
 #define MAJOR_NUM 235
-#define MSG_SLOT_CHANNEL _IOW(MAJOR_NUM, 0, unsigned long)
+#define MSG_SLOT_CHANNEL _IOW(MAJOR_NUM, 0, unsigned int)
 #define DEVICE_RANGE_NAME "message_slot"
 #define BUF_LEN 128
 #define SUCCESS 0
